removing utils - can just be added to main train script
