+++ train.py
@@ -1,17 +1,41 @@
 import os
 import time
+
 import wandb
 from datasets import load_dataset
-from transformers import TrainingArguments, Trainer, AutoTokenizer
-from transformers import OPTConfig, OPTForCausalLM
-from src.utils.utils import get_deepspeed_config
+from transformers import (
+    AutoTokenizer,
+    OPTConfig,
+    OPTForCausalLM,
+    Trainer,
+    TrainingArguments,
+)
 
-
+# Constants and Helpers
 LOG_EVERY_N_STEPS = 10
-SAVE_EVERY_N_STEPS = 500 
+SAVE_EVERY_N_STEPS = 500
 TRAIN_EPOCHS = 10
 
-def train_model(model_type="opt", seq_len=128, use_deepspeed=False, push_to_hub=True, dry_run=False):
+
+def get_deepspeed_config():
+    return {
+        "zero_optimization": {
+            "stage": 3,
+            "offload_optimizer": {"device": "cpu"},
+            "offload_param": {"device": "cpu"},
+        },
+        "train_batch_size": 64,
+        "gradient_accumulation_steps": 1,
+        "fp16": {"enabled": True},
+    }
+
+
+# Model Train Script
+
+
+def train_model(
+    model_type="opt", seq_len=128, use_deepspeed=False, push_to_hub=True, dry_run=False
+):
     dataset = load_dataset(f"babylm-seqlen/train_100M_{seq_len}_single_shuffle")
     dataset = dataset.map(lambda x: {"labels": x["input_ids"]})
 
@@ -44,6 +68,7 @@ def train_model(model_type="opt", seq_len=128, use_deepspeed=False, push_to_hub=
 
     elif model_type == "mamba":
         from transformers import MambaConfig, MambaForCausalLM
+
         config = MambaConfig(
             vocab_size=50257,
             hidden_size=256,
@@ -65,8 +90,8 @@ def train_model(model_type="opt", seq_len=128, use_deepspeed=False, push_to_hub=
 
     training_args = TrainingArguments(
         output_dir=output_dir,
-        per_device_train_batch_size=64, # TODO: tune this
-        gradient_accumulation_steps=4, # TODO: tune this
+        per_device_train_batch_size=64,  # TODO: tune this
+        gradient_accumulation_steps=4,  # TODO: tune this
         num_train_epochs=TRAIN_EPOCHS,
         evaluation_strategy="no",
         save_strategy="steps",
@@ -93,16 +118,25 @@ def train_model(model_type="opt", seq_len=128, use_deepspeed=False, push_to_hub=
     trainer.train()
     end_time = time.time()
 
-    print(f"✅ Training {model_type.upper()} for seq_len {seq_len} done in {end_time - start_time:.2f}s")
+    print(
+        f"✅ Training {model_type.upper()} for seq_len {seq_len} done in {end_time - start_time:.2f}s"
+    )
 
 
 if __name__ == "__main__":
     import argparse
+
     parser = argparse.ArgumentParser()
-    parser.add_argument("--model_type", type=str, default="opt", choices=["opt", "mamba"])
+    parser.add_argument(
+        "--model_type", type=str, default="opt", choices=["opt", "mamba"]
+    )
     parser.add_argument("--seq_len", type=int, default=128)
     parser.add_argument("--use_deepspeed", action="store_true")
-    parser.add_argument("--no_push_to_hub", action="store_true", help="If set, do NOT push to the Hugging Face Hub.")
+    parser.add_argument(
+        "--no_push_to_hub",
+        action="store_true",
+        help="If set, do NOT push to the Hugging Face Hub.",
+    )
     parser.add_argument("--dry_run", action="store_true")
 
     args = parser.parse_args()
@@ -112,5 +146,5 @@ if __name__ == "__main__":
         seq_len=args.seq_len,
         use_deepspeed=args.use_deepspeed,
         push_to_hub=not args.no_push_to_hub,
-        dry_run=args.dry_run
+        dry_run=args.dry_run,
     )
